Fixed, I believe, all API inconsistencies wrt namings, both CLI commands and internal methods

+++ Threadle.Core/Analysis/Analyses.cs
@@ -238,7 +238,7 @@ namespace Threadle.Core.Analysis
                 return OperationResult.Fail("InvalidLayerType", $"Layer '{layerName}' is neither a one-mode nor a two-mode layer.");
             // Convert the degree mapping into the <uint,string> format and store this as a node attribute
             var attrDict = degreeMapping.ToDictionary(kvp => kvp.Key, kvp => kvp.Value.ToString());
-            return network.Nodeset.DefineAndSetNodeAttributeValues(attrName, attrDict, NodeAttributeType.Float);
+            return network.Nodeset.DefineAndSetNodeAttributeValues(attrName, attrDict, NodeAttributeType.Int);
         }
 
         /// <summary>
